Increased test coverage and fixed a potential rounding problem
The previous implementation always returned zero for any era lasting less than 2.5 minutes.

After this change it works for smaller eras as well.

There may still be minor rounding issues during the division, but it's fine as this payout mechanism is temporary and works for era times that are multiple of 6 seconds (current block time), which is a divisor of 150 seconds (Zend block time).