Added CI workflow for Rust doc generation
The workflow automatically generates the Rust documentation and publishes it directly to GitHub pages.

To make navigation easier, since there is no default index.html, the CI automatically provides one containing just a redirect to the index of the nh-node crate.

Note: the workflow is currently meant to be triggered manually or by another workflow.